+++ packages/backend/src/types.ts
@@ -26,6 +26,7 @@ import {
 import { PluginTaskScheduler } from '@backstage/backend-tasks';
 import { IdentityApi } from '@backstage/plugin-auth-node';
 import { PermissionEvaluator } from '@backstage/plugin-permission-common';
+import { EventBroker } from '@backstage/plugin-events-node';
 
 export type PluginEnvironment = {
   logger: Logger;
@@ -38,4 +39,5 @@ export type PluginEnvironment = {
   permissions: PermissionEvaluator;
   scheduler: PluginTaskScheduler;
   identity: IdentityApi;
+  eventBroker: EventBroker;
 };