+++ plugins/kubernetes-backend/src/service/KubernetesFanOutHandler.test.ts
@@ -166,6 +166,11 @@ function getKubernetesFanOutHandler(customResources: CustomResource[]) {
       getClustersByEntity,
     },
     customResources: customResources,
+    authTranslator: {
+      decorateClusterDetailsWithAuth: async (clusterDetails, _) => {
+        return clusterDetails;
+      },
+    },
   });
 }
 
@@ -879,6 +884,11 @@ describe('getKubernetesObjectsByEntity', () => {
             objectType: 'services',
           },
         ],
+        authTranslator: {
+          decorateClusterDetailsWithAuth: async (clusterDetails, _) => {
+            return clusterDetails;
+          },
+        },
       });
 
       const result = await sut.getKubernetesObjectsByEntity({