+++ plugins/kubernetes-backend/src/service/KubernetesProxy.test.ts
@@ -28,16 +28,19 @@ import { ClusterDetails, KubernetesClustersSupplier } from '../types/types';
 import {
   APPLICATION_JSON,
   HEADER_KUBERNETES_CLUSTER,
+  HEADER_KUBERNETES_AUTH,
   KubernetesProxy,
 } from './KubernetesProxy';
 import {
   AuthorizeResult,
   PermissionEvaluator,
 } from '@backstage/plugin-permission-common';
+import { KubernetesAuthTranslator } from '../kubernetes-auth-translator';
 
 describe('KubernetesProxy', () => {
   let proxy: KubernetesProxy;
   const worker = setupServer();
+  const logger = getVoidLogger();
 
   setupRequestMockHandlers(worker);
 
@@ -73,9 +76,13 @@ describe('KubernetesProxy', () => {
     authorizeConditional: jest.fn(),
   };
 
+  const authTranslator: jest.Mocked<KubernetesAuthTranslator> = {
+    decorateClusterDetailsWithAuth: jest.fn(),
+  };
+
   beforeEach(() => {
     jest.resetAllMocks();
-    proxy = new KubernetesProxy(getVoidLogger(), clusterSupplier);
+    proxy = new KubernetesProxy({ logger, clusterSupplier, authTranslator });
   });
 
   it('should return a ERROR_NOT_FOUND if no clusters are found', async () => {
@@ -112,26 +119,219 @@ describe('KubernetesProxy', () => {
         authProvider: 'serviceAccount',
       },
     ] as ClusterDetails[]);
+
+    permissionApi.authorize.mockReturnValue(
+      Promise.resolve([{ result: AuthorizeResult.ALLOW }]),
+    );
+
+    authTranslator.decorateClusterDetailsWithAuth.mockResolvedValue({
+      name: 'cluster1',
+      url: 'https://localhost:9999',
+      serviceAccountToken: '',
+      authProvider: 'serviceAccount',
+    } as ClusterDetails);
+
     const app = express().use(
       '/mountpath',
       proxy.createRequestHandler({ permissionApi }),
-    );
-    permissionApi.authorize.mockReturnValue(
-      Promise.resolve([{ result: AuthorizeResult.ALLOW }]),
     );
     const requestPromise = request(app)
       .get('/mountpath/api')
       .set(HEADER_KUBERNETES_CLUSTER, 'cluster1');
     worker.use(
-      rest.get('https://localhost:9999/api', (_, res, ctx) =>
+      rest.get('https://localhost:9999/api', (_: any, res: any, ctx: any) =>
         res(ctx.status(299), ctx.json(apiResponse)),
       ),
-      rest.all(requestPromise.url, (req, _res, _ctx) => req.passthrough()),
+      rest.all(requestPromise.url, (req: any) => req.passthrough()),
     );
 
     const response = await requestPromise;
 
     expect(response.status).toEqual(299);
     expect(response.body).toStrictEqual(apiResponse);
+  });
+
+  it('should default to using a provided authorization header', async () => {
+    worker.use(
+      rest.get(
+        'https://localhost:9999/api/v1/namespaces',
+        (req: any, res: any, ctx: any) => {
+          if (!req.headers.get('Authorization')) {
+            return res(ctx.status(401));
+          }
+
+          if (req.headers.get('Authorization') !== 'my-token') {
+            return res(ctx.status(403));
+          }
+
+          return res(
+            ctx.status(200),
+            ctx.json({
+              kind: 'NamespaceList',
+              apiVersion: 'v1',
+              items: [],
+            }),
+          );
+        },
+      ),
+    );
+
+    permissionApi.authorize.mockReturnValue(
+      Promise.resolve([{ result: AuthorizeResult.ALLOW }]),
+    );
+
+    clusterSupplier.getClusters.mockResolvedValue([
+      {
+        name: 'cluster1',
+        url: 'https://localhost:9999',
+        serviceAccountToken: '',
+        authProvider: 'serviceAccount',
+      },
+    ] as ClusterDetails[]);
+
+    authTranslator.decorateClusterDetailsWithAuth.mockResolvedValue({
+      name: 'cluster1',
+      url: 'https://localhost:9999',
+      serviceAccountToken: 'random-token',
+      authProvider: 'serviceAccount',
+    } as ClusterDetails);
+
+    const app = express().use(
+      '/mountpath',
+      proxy.createRequestHandler({ permissionApi }),
+    );
+    const requestPromise = request(app)
+      .get('/mountpath/api/v1/namespaces')
+      .set(HEADER_KUBERNETES_CLUSTER, 'cluster1')
+      .set('Authorization', 'my-token');
+
+    worker.use(rest.all(requestPromise.url, (req: any) => req.passthrough()));
+
+    const response = await requestPromise;
+
+    expect(response.status).toEqual(200);
+  });
+
+  it('should add a serviceAccountToken to the request headers if one isnt provided in request and one isnt set up in cluster details', async () => {
+    worker.use(
+      rest.get('https://localhost:9999/api/v1/namespaces', (req, res, ctx) => {
+        if (!req.headers.get('Authorization')) {
+          return res(ctx.status(401));
+        }
+
+        if (req.headers.get('Authorization') !== 'Bearer my-token') {
+          return res(ctx.status(403));
+        }
+
+        return res(
+          ctx.status(200),
+          ctx.json({
+            kind: 'NamespaceList',
+            apiVersion: 'v1',
+            items: [],
+          }),
+        );
+      }),
+    );
+
+    permissionApi.authorize.mockReturnValue(
+      Promise.resolve([{ result: AuthorizeResult.ALLOW }]),
+    );
+
+    clusterSupplier.getClusters.mockResolvedValue([
+      {
+        name: 'cluster1',
+        url: 'https://localhost:9999',
+        authProvider: 'googleServiceAccount',
+      },
+    ] as ClusterDetails[]);
+
+    authTranslator.decorateClusterDetailsWithAuth.mockResolvedValue({
+      name: 'cluster1',
+      url: 'https://localhost:9999',
+      serviceAccountToken: 'my-token',
+      authProvider: 'googleServiceAccount',
+    } as ClusterDetails);
+
+    const app = express().use(
+      '/mountpath',
+      proxy.createRequestHandler({ permissionApi }),
+    );
+    const requestPromise = request(app)
+      .get('/mountpath/api/v1/namespaces')
+      .set(HEADER_KUBERNETES_CLUSTER, 'cluster1');
+
+    worker.use(rest.all(requestPromise.url, (req: any) => req.passthrough()));
+
+    const response = await requestPromise;
+
+    expect(response.status).toEqual(200);
+    expect(response.body).toStrictEqual({
+      kind: 'NamespaceList',
+      apiVersion: 'v1',
+      items: [],
+    });
+  });
+
+  it('should append the Backstage-Kubernetes-Auth field to the requests authorization header if one is provided', async () => {
+    worker.use(
+      rest.get('https://localhost:9999/api/v1/namespaces', (req, res, ctx) => {
+        if (!req.headers.get('Authorization')) {
+          return res(ctx.status(401));
+        }
+
+        if (req.headers.get('Authorization') !== 'tokenB') {
+          return res(ctx.status(403));
+        }
+
+        return res(
+          ctx.status(200),
+          ctx.json({
+            kind: 'NamespaceList',
+            apiVersion: 'v1',
+            items: [],
+          }),
+        );
+      }),
+    );
+
+    permissionApi.authorize.mockReturnValue(
+      Promise.resolve([{ result: AuthorizeResult.ALLOW }]),
+    );
+
+    clusterSupplier.getClusters.mockResolvedValue([
+      {
+        name: 'cluster1',
+        url: 'https://localhost:9999',
+        authProvider: 'googleServiceAccount',
+      },
+    ] as ClusterDetails[]);
+
+    authTranslator.decorateClusterDetailsWithAuth.mockResolvedValue({
+      name: 'cluster1',
+      url: 'https://localhost:9999',
+      serviceAccountToken: 'tokenA',
+      authProvider: 'googleServiceAccount',
+    } as ClusterDetails);
+
+    const app = express().use(
+      '/mountpath',
+      proxy.createRequestHandler({ permissionApi }),
+    );
+    const requestPromise = request(app)
+      .get('/mountpath/api/v1/namespaces')
+      .set(HEADER_KUBERNETES_CLUSTER, 'cluster1')
+      .set(HEADER_KUBERNETES_AUTH, 'tokenB');
+
+    worker.use(rest.all(requestPromise.url, (req: any) => req.passthrough()));
+
+    const response = await requestPromise;
+
+    expect(response.status).toEqual(200);
+    expect(response.body).toStrictEqual({
+      kind: 'NamespaceList',
+      apiVersion: 'v1',
+      items: [],
+    });
   });
 });