+++ packages/core-app-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/core-app-api
 
+## 1.7.0-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/config@1.0.7
+  - @backstage/types@1.0.2
+  - @backstage/version-bridge@1.0.4-next.0
+
 ## 1.7.0-next.1
 
 ### Patch Changes