+++ packages/e2e-test/CHANGELOG.md
@@ -1,5 +1,14 @@
 # e2e-test
 
+## 0.2.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/create-app@0.4.39-next.2
+  - @backstage/cli-common@0.1.12
+  - @backstage/errors@1.1.5
+
 ## 0.2.2-next.1
 
 ### Patch Changes