+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,24 @@
 # techdocs-cli-embedded-app
 
+## 0.2.81-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli@0.22.6-next.2
+  - @backstage/core-components@0.12.6-next.2
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/plugin-techdocs@1.6.1-next.2
+  - @backstage/app-defaults@1.3.0-next.2
+  - @backstage/catalog-model@1.2.1
+  - @backstage/config@1.0.7
+  - @backstage/core-app-api@1.7.0-next.2
+  - @backstage/integration-react@1.1.12-next.2
+  - @backstage/test-utils@1.3.0-next.2
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/plugin-catalog@1.10.0-next.2
+  - @backstage/plugin-techdocs-react@1.1.5-next.2
+
 ## 0.2.81-next.1
 
 ### Patch Changes