+++ plugins/catalog-customized/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @internal/plugin-catalog-customized
 
+## 0.0.9-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.4.1-next.2
+  - @backstage/plugin-catalog@1.10.0-next.2
+
 ## 0.0.9-next.1
 
 ### Patch Changes