+++ plugins/kubernetes/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-kubernetes
 
+## 0.7.10-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.6-next.2
+  - @backstage/plugin-catalog-react@1.4.1-next.2
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/catalog-model@1.2.1
+  - @backstage/config@1.0.7
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/plugin-kubernetes-common@0.6.2-next.1
+
 ## 0.7.10-next.1
 
 ### Patch Changes